requirements.txt
streamlit
typing_extensions
langgraph
pypdf
langchain_groq
langchain_core
tavily-python
groq
python-dotenv
scikit-learn
python-docx
Pillow
pytesseract
